turing/requirements.txt
asgiref==3.11.1
beautifulsoup4==4.12.3
crispy-bootstrap4==2026.2
Django==6.0.4
django-crispy-forms==2.6
django-simple-history==3.11.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2026.1.post1
requests==2.32.3
six==1.17.0
sqlparse==0.5.5
whitenoise==6.12.0
